create interface for adapters and add implementations (#4824)
This PR adds a GraphQLEngineAdapter interface which is implemented by
`gqlengineadapter.ProxyOnly`, `gqlengineadapter.Supergraph` and
`gqlengineadapter.UniversalDataGraph`.

This also means that those implementations moved to a subpackage called
`gqlengineadapter`.
The `adapter` package is meant to be a general package for API
definition adapters (like AsyncAPI), so it makes sense not to clutter it
with GraphQL Engine Adapters.

## Types of changes

<!-- What types of changes does your code introduce? Put an `x` in all
the boxes that apply: -->

- [ ] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to change)
- [x] Refactoring or add test (improvements in base code or adds test
coverage to functionality)
