+++ apidef/adapter/gqlengineadapter/utils_engine_v2.go
@@ -1,4 +1,4 @@
-package adapter
+package gqlengineadapter
 
 import (
 	"errors"
@@ -134,4 +134,11 @@ func createGraphQLDataSourceFactory(params createGraphQLDataSourceFactoryParams)
 	}
 	factory.SubscriptionClient = subscriptionClient
 	return factory, nil
+}
+
+func subscriptionClientFactoryOrDefault(providedSubscriptionClientFactory graphqlDataSource.GraphQLSubscriptionClientFactory) graphqlDataSource.GraphQLSubscriptionClientFactory {
+	if providedSubscriptionClientFactory != nil {
+		return providedSubscriptionClientFactory
+	}
+	return &graphqlDataSource.DefaultSubscriptionClientFactory{}
 }
